timestamps: rename nc-multiplex-sri to nc-multiplex, update startup scripts
